requirements.txt
# MediScan AI - Python Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Request parsing
python-multipart==0.0.9
pydantic[email]==2.9.2

# ML / Inference
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0

# Training only (Kaggle notebook)
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv==1.0.1
